# Issue #595 faster Translation Memory (tm) API