--- a/app/classes/Transvision/ShowResults.php
+++ b/app/classes/Transvision/ShowResults.php
@@ -93,16 +93,18 @@ class ShowResults
      * Return an array of search results from our Translation Memory API
      * service with a quality index based on the levenshtein distance.
      *
-     * @param  array  $entities      The entities we want to analyse
-     * @param  array  $array_strings The strings to look into [locale1 strings, locale2 strings]
-     * @param  string $search        The string to search for
-     * @param  int    $max_results   Optional, default to 200, the max number of results we return
-     * @param  int    $min_quality   Optional, default to 0, The minimal quality index to filter result
+     * @param  array  $source_strings The source reference strings with entities as keys
+     * @param  array  $target_strings The target strings to look into with entities as keys
+     * @param  string $search         The string to search for
+     * @param  int    $max_results    Optional, default to 200, the max number of results we return
+     * @param  int    $min_quality    Optional, default to 0, The minimal quality index to filter result
      * @return array  An array of strings as [source => string, target => string, quality=> Levenshtein index]
      */
-    public static function getTranslationMemoryResults($entities, $array_strings, $search, $max_results = 200, $min_quality = 0)
+    public static function getTranslationMemoryResults($source_strings, $target_strings, $search, $max_results = 200, $min_quality = 0)
     {
-        $search_results = array_values(self::getTMXResults($entities, $array_strings));
+        $search_results = array_values(
+            self::getTMXResults(array_keys($source_strings), [$source_strings, $target_strings])
+        );
         $output = [];
 
         foreach ($search_results as $set) {