```
/api/v1/tm/global/en-US/fr/Cancel/
```
Before:
[Wed Jan 20 00:04:54 2016] Memory peak: 23592960 (22.5MB)
[Wed Jan 20 00:04:54 2016] Elapsed time (s): 0.3324

After:
[Wed Jan 20 04:12:04 2016] Memory peak: 8388608 (8MB)
[Wed Jan 20 04:12:04 2016] Elapsed time (s): 0.145

The gain is ~x2.2

If the search is for a string that exists only in one small repository, mozilla.org for example, the gains are even bigger (~x4):

```
/api/v1/tm/global/en-US/fr/Foundation pledge/
```
Before:
[Wed Jan 20 01:04:55 2016] Memory peak: 23330816 (22.25MB)
[Wed Jan 20 01:04:55 2016] Elapsed time (s): 0.2898

After:
Wed Jan 20 04:14:04 2016] Memory peak: 7077888 (6.75MB)
[Wed Jan 20 04:14:04 2016] Elapsed time (s): 0.0796